requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.1
pillow==10.2.0
psycopg2==2.9.6
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
